requirements.txt
numpy
fastapi
uvicorn[standard]
pydantic
requests
pillow
opencv-python
easyocr
python-multipart